Add style.rb to lib
