+++ lib/messages.rb
@@ -1,7 +1,9 @@
 require 'colorize'
+require_relative '../lib/style_rules'
+require_relative '../lib/style'
+require_relative 'main.rb'
 
 module Messages
-  DONE = '              LINTER HAS FINISHED CHECKING YOUR FILE'.bold
-  LINE = '              **************************************'
+  DONE = "Linter has finished checking your file and #{errors.length} found".bold
   NO_ERRORS = 'No errors'.green.bold
 end
