+++ test.css
@@ -1,4 +1,4 @@
-.test {
+.test{
   background: #fff;
   margin: 0;
 }
